+++ app/templates/components/public/side-menu-inner.hbs
@@ -1,0 +1,128 @@
+{{#if (eq this.session.currentRouteName 'public.index')}}
+  <a href='#info' {{action "scrollToTarget" 'info' }} class='item active scroll'>
+    <div>
+      <i class="info icon"></i>
+      {{t 'Info'}}
+    </div>
+  </a>
+  {{#if this.event.tickets.length}}
+    <a href='#tickets' {{action "scrollToTarget" 'tickets' }} class='item scroll'>
+      <div>
+        <i class="ticket icon"></i>
+        {{t 'Tickets'}}
+      </div>
+    </a>
+  {{/if}}
+  {{#if this.showSpeakers}}
+    <LinkTo @route="public.speakers" class="item">
+      <div>
+        <i class="bullhorn icon"></i>
+        {{t 'Speakers'}}
+      </div>
+    </LinkTo>
+   {{/if}}
+{{else}}
+  <LinkTo class="item" @route="public.index" {{action 'goToSection' 'info' }}>
+    <div>
+      <i class="info icon"></i>
+      {{t 'Info'}}
+    </div>
+  </LinkTo>
+  {{#if this.event.tickets.length}}
+    <LinkTo class="item" @route="public.index" {{action 'goToSection' 'tickets' }}>
+      <div>
+        <i class="ticket icon"></i>
+        {{t 'Tickets'}}
+      </div>
+    </LinkTo>
+  {{/if}}
+  {{#if this.showSpeakers}}
+    <LinkTo @route="public.speakers" class="item">
+      <div>
+        <i class="bullhorn icon"></i>
+        {{t 'Speakers'}}
+      </div>
+    </LinkTo>
+  {{/if}}
+{{/if}}
+{{#if (and this.isSchedulePublished this.showSessions)}}
+  <LinkTo @route="public.sessions" class="item">
+    <div>
+      <i class="book icon"></i>
+      {{t 'Schedule'}}
+    </div>
+  </LinkTo>
+{{/if}}
+{{#if (and this.isSchedulePublished this.showSessions)}}
+  <LinkTo @route="public.schedule" class="item">
+    <div>
+      <i class="calendar icon"></i>
+      {{t 'Calendar'}}
+    </div>
+  </LinkTo>
+{{/if}}
+{{#if this.shouldShowCallforSpeakers }}
+  <LinkTo @route="public.cfs" class="item">
+    <div>
+      <i class="bullhorn icon"></i>
+      {{t 'Call for Speakers'}}
+    </div>
+  </LinkTo>
+{{/if}}
+{{#if (and this.event.isSponsorsEnabled this.event.sponsors)}}
+ {{#if (eq this.session.currentRouteName 'public.index')}}
+   <a href='#sponsor' {{action "scrollToTarget" 'sponsor' }} class='item scroll'>
+    <div>
+      <i class="dollar icon"></i>
+      {{t 'Sponsors'}}
+    </div>
+   </a>
+ {{else}}
+   <LinkTo class="item" @route="public.index" {{action 'goToSection' 'sponsor' }}>
+      <div>
+        <i class="dollar icon"></i>
+        {{t 'Sponsors'}}
+      </div>
+   </LinkTo>
+ {{/if}}
+{{/if}}
+{{#if this.event.hasOwnerInfo}}
+ {{#if (eq this.session.currentRouteName 'public.index')}}
+   <a href='#organizer' {{action "scrollToTarget" 'organizer' }} class='item scroll'>
+    <div>
+      <i class="users icon"></i>
+      {{t 'Organizer'}}
+    </div>
+   </a>
+ {{else}}
+   <LinkTo class="item" @route="public.index" {{action 'goToSection' 'organizer' }}>
+    <div>
+      <i class="users icon"></i>
+      {{t 'Organizer'}}
+    </div>
+   </LinkTo>
+ {{/if}}
+{{/if}}
+{{#if (eq this.session.currentRouteName 'public.index')}}
+  <a href='#getting-here' {{action "scrollToTarget" 'getting-here' }} class='item scroll'>
+    <div>
+      <i class="icon map marker alternate"></i>
+      {{t 'Getting here'}}
+    </div>
+  </a>
+{{else}}
+  <LinkTo class="item" @route="public.index" {{action 'goToSection' 'getting-here' }}>
+    <div>
+      <i class="icon map marker alternate"></i>
+      {{t 'Getting here'}}
+    </div>
+  </LinkTo>
+{{/if}}
+{{#if this.event.codeOfConduct}}
+  <LinkTo @route="public.coc" class="item">
+    <div>
+      <i class="book icon"></i>
+      {{t 'Code of Conduct'}}
+    </div>
+  </LinkTo>
+{{/if}}
