+++ app/templates/components/public/side-menu.hbs
@@ -2,190 +2,16 @@
   {{#if this.device.isMobile}}
     <UiDropdown class="ui selection dropdown fluid vertical" @selected={{this.activeMenuSection}}>
       <div class="default text">
-          {{this.activeMenu}}
+        {{this.activeMenu}}
       </div>
       <i class="dropdown icon"></i>
       <div style="max-height: none !important;" class="menu">
-        {{#if (eq this.session.currentRouteName 'public.index')}}
-          <a href='#info' {{action "scrollToTarget" 'info'}} class='active item scroll'>
-            <i class="info icon"></i>
-            {{t 'Info'}}
-          </a>
-          {{#if this.event.tickets.length}}
-            <a href='#tickets' {{action "scrollToTarget" 'tickets'}} class='item scroll'>
-              <i class="ticket icon"></i>
-              {{t 'Tickets'}}
-            </a>
-          {{/if}}
-          {{#if this.showSpeakers}}
-            <LinkTo @route="public.speakers" class="item">
-              <i class="bullhorn icon"></i>
-              {{t 'Speakers'}}
-            </LinkTo>
-          {{/if}}
-        {{else}}
-            <LinkTo class="item" @route="public.index" {{action 'goToSection' 'info'}}>
-              <i class="info icon"></i>
-              {{t 'Info'}}
-            </LinkTo>
-            {{#if this.event.tickets.length}}
-              <LinkTo class="item" @route="public.index" {{action 'goToSection' 'tickets'}}>
-              <i class="ticket icon"></i>
-                {{t 'Tickets'}}
-              </LinkTo>
-            {{/if}}
-            {{#if this.showSpeakers}}
-              <LinkTo @route="public.speakers" class="item">
-                <i class="bullhorn icon"></i>
-                {{t 'Speakers'}}
-              </LinkTo>
-            {{/if}}
-        {{/if}}
-        {{#if (and this.isSchedulePublished this.showSessions)}}
-          <LinkTo @route="public.sessions" class="item">
-            <i class="book icon"></i>
-            {{t 'Schedule'}}
-          </LinkTo>
-        {{/if}}
-        {{#if (and this.isSchedulePublished this.showSessions)}}
-          <LinkTo @route="public.schedule" class="item">
-            <i class="calendar icon"></i>
-            {{t 'Calendar'}}
-          </LinkTo>
-        {{/if}}  
-        {{#if this.shouldShowCallforSpeakers }}
-          <LinkTo @route="public.cfs" class="item">
-            <i class="bullhorn icon"></i>
-            {{t 'Call for Speakers'}}
-          </LinkTo>
-        {{/if}}
-        {{#if (and this.event.isSponsorsEnabled this.event.sponsors)}}
-          {{#if (eq this.session.currentRouteName 'public.index')}}
-            <a href='#sponsor' {{action "scrollToTarget" 'sponsor'}} class='item scroll'>
-              <i class="dollar icon"></i>
-              {{t 'Sponsors'}}
-            </a>
-          {{else}}
-            <LinkTo class="item" @route="public.index" {{action 'goToSection' 'sponsor'}}>
-              <i class="dollar icon"></i>
-              {{t 'Sponsors'}}
-            </LinkTo>
-          {{/if}}
-        {{/if}}
-        {{#if this.event.hasOwnerInfo}}
-          {{#if (eq this.session.currentRouteName 'public.index')}}
-            <a href='#organizer' {{action "scrollToTarget" 'organizer'}} class='item scroll'>
-              <i class="users icon"></i>
-              {{t 'Organizer'}}
-            </a>
-          {{else}}
-            <LinkTo class="item" @route="public.index" {{action 'goToSection' 'organizer'}}>
-              <i class="users icon"></i>
-              {{t 'Organizer'}}
-            </LinkTo>
-          {{/if}}
-        {{/if}}
-        {{#if (eq this.session.currentRouteName 'public.index')}}
-          <a href='#getting-here' {{action "scrollToTarget" 'getting-here'}} class='item scroll'>
-            <i class="icon map marker alternate"></i>
-            {{t 'Getting here'}}
-          </a>
-        {{else}}
-          <LinkTo class="item" @route="public.index" {{action 'goToSection' 'getting-here'}}>
-            <i class="icon map marker alternate"></i>
-            {{t 'Getting here'}}
-          </LinkTo>
-        {{/if}}
-        {{#if this.event.codeOfConduct}}
-          <LinkTo @route="public.coc" class="item">
-            <i class="book icon"></i>
-            {{t 'Code of Conduct'}}
-          </LinkTo>
-        {{/if}} 
+        <Public::SideMenuInner @activeMenuSection={{this.activeMenuSection}} @activeSection={{this.activeSection}} @event={{@event}} />
       </div>
     </UiDropdown>
   {{else}}
     <div class="ui fluid vertical pointing menu">
-      {{#if (eq this.session.currentRouteName 'public.index')}}
-        <a href='#info' {{action "scrollToTarget" 'info'}} class='item active scroll'>
-          {{t 'Info'}}
-        </a>
-        {{#if this.event.tickets.length}}
-          <a href='#tickets' {{action "scrollToTarget" 'tickets'}} class='item scroll'>
-            {{t 'Tickets'}}
-          </a>
-        {{/if}}
-        {{#if this.showSpeakers}}
-          <LinkTo @route="public.speakers" class="item">
-            {{t 'Speakers'}}
-          </LinkTo>
-        {{/if}}
-      {{else}}
-          <LinkTo class="item" @route="public.index" {{action 'goToSection' 'info'}}>
-            {{t 'Info'}}
-          </LinkTo>
-        {{#if this.event.tickets.length}}
-          <LinkTo class="item" @route="public.index" {{action 'goToSection' 'tickets'}}>
-            {{t 'Tickets'}}
-          </LinkTo>
-        {{/if}}
-        {{#if this.showSpeakers}}
-          <LinkTo @route="public.speakers" class="item">
-            {{t 'Speakers'}}
-          </LinkTo>
-        {{/if}}
-      {{/if}}
-      {{#if (and this.isSchedulePublished this.showSessions)}}
-        <LinkTo @route="public.sessions" class="item">
-          {{t 'Schedule'}}
-        </LinkTo>
-      {{/if}}
-      {{#if (and this.isSchedulePublished this.showSessions)}}
-        <LinkTo @route="public.schedule" class="item">
-          {{t 'Calendar'}}
-        </LinkTo>
-      {{/if}}  
-      {{#if this.shouldShowCallforSpeakers }}
-        <LinkTo @route="public.cfs" class="item">
-          {{t 'Call for Speakers'}}
-        </LinkTo>
-      {{/if}}
-      {{#if (and this.event.isSponsorsEnabled this.event.sponsors)}}
-        {{#if (eq this.session.currentRouteName 'public.index')}}
-          <a href='#sponsor' {{action "scrollToTarget" 'sponsor'}} class='item scroll'>
-            {{t 'Sponsors'}}
-          </a>
-        {{else}}
-          <LinkTo class="item" @route="public.index" {{action 'goToSection' 'sponsor'}}>
-            {{t 'Sponsors'}}
-          </LinkTo>
-        {{/if}}
-      {{/if}}
-      {{#if this.event.hasOwnerInfo}}
-        {{#if (eq this.session.currentRouteName 'public.index')}}
-          <a href='#organizer' {{action "scrollToTarget" 'organizer'}} class='item scroll'>
-            {{t 'Organizer'}}
-          </a>
-        {{else}}
-          <LinkTo class="item" @route="public.index" {{action 'goToSection' 'organizer'}}>
-            {{t 'Organizer'}}
-          </LinkTo>
-        {{/if}}
-      {{/if}}
-      {{#if (eq this.session.currentRouteName 'public.index')}}
-        <a href='#getting-here' {{action "scrollToTarget" 'getting-here'}} class='item scroll'>
-          {{t 'Getting here'}}
-        </a>
-      {{else}}
-        <LinkTo class="item" @route="public.index" {{action 'goToSection' 'getting-here'}}>
-          {{t 'Getting here'}}
-        </LinkTo>
-      {{/if}}
-      {{#if this.event.codeOfConduct}}
-        <LinkTo @route="public.coc" class="item">
-          {{t 'Code of Conduct'}}
-        </LinkTo>
-      {{/if}}
+      <Public::SideMenuInner @activeMenuSection={{this.activeMenuSection}} @activeSection={{this.activeSection}} @event={{@event}} />
     </div>
   {{/if}}
 {{/if}}
